fix(server-dev): Simplify server-dev manager code.
