+++ .swcrc
@@ -4,9 +4,10 @@
     "parser": {
       "syntax": "ecmascript",
       "jsx": true,
-      "dynamicImport": true
+      "dynamicImport": true,
+      "topLevelAwait": true
     },
-    "target": "es2020",
+    "target": "es2022",
     "keepClassNames": true,
     "transform": { "react": { "runtime": "classic" } }
   },
