+++ packages/server-dev/manager/run.mjs
@@ -79,27 +79,26 @@ when:
 - initial build fails
 */
 
-async function run() {
-  const context = await getContext();
+const context = await getContext();
+
+try {
   try {
-    try {
-      await context.initialBuild();
-    } catch (error) {
-      context.logger.error(error);
-    }
-    await context.startWatchers();
-    startServer(context);
-    await wait(800);
-    if (process.env.LOWDEFY_SERVER_DEV_OPEN_BROWSER === 'true') {
-      // TODO: Wait 1 sec for a ping and don't open if a ping is seen
-      opener(`http://localhost:${context.options.port}`);
-    }
-    await new Promise(() => {});
+    await context.initialBuild();
   } catch (error) {
     context.logger.error(error);
-    context.shutdownServer();
-    process.exit();
   }
-}
 
-run();
+  await context.startWatchers();
+
+  startServer(context);
+  await wait(800);
+  if (process.env.LOWDEFY_SERVER_DEV_OPEN_BROWSER === 'true') {
+    // TODO: Wait 1 sec for a ping and don't open if a ping is seen
+    opener(`http://localhost:${context.options.port}`);
+  }
+  await new Promise(() => {});
+} catch (error) {
+  context.logger.error(error);
+  context.shutdownServer();
+  process.exit();
+}
